fix: extract health probes to lib/ so next build accepts route.ts
Next.js 14 rejects arbitrary named exports from route.ts files —
only HTTP method handlers + a small set of config constants (runtime,
dynamic, revalidate, etc.) are allowed. My initial M11-7 put
checkBudgetResetBacklog + checkSupabase directly in app/api/health/
route.ts, which vitest accepted (test job passed) but `next build`
did not (build job failed on PR #94).

Fix: move both probes to lib/health-checks.ts. Route file now holds
only the GET handler + imports. Test file imports from the new lib
module. Zero behavioural change — probe logic, response shape,
degraded-status semantics, and 503 threshold are all identical.

Verified locally: typecheck / lint / next build all clean.
